+++ src/js/src/plotly/constants.js
@@ -32,8 +32,6 @@ module.exports = {
     svg: /^data:image\/svg\+xml,/
   },
 
-  mathJaxConfigQuery: '?config=TeX-AMS-MML_SVG',
-
   // time [in ms] after which printToPDF errors when image isn't loaded
   pdfPageLoadImgTimeout: 20000
 }